use ffmpeg.wasm to synthesize two files
The download function is rewritten, and the desired function is to automatically use ffmpeg.wasm to synthesize video and audio and output a synthesized file, but it is limited by the cross-domain policy of the website. If it is the Chrome browser, you need to use the command line flag --enable-features=SharedArrayBuffer to start Chrome, which is currently the case and cannot be solved.